fix: return correct id on upsert
The `last_insert_rowid()` function in `rusqlite` returns 0 when an
`INSERT ... ON CONFLICT ... DO UPDATE` statement results in an update.
This was causing tests to fail because they were expecting the actual
row id.

This commit changes the `upsert` functions in `Entry` and `Attribute`
to fetch the row after the upsert operation to ensure the correct id is
returned.